fix: remove unnecessary polyfills (#5)
removes some unnecessary polyfills when building CSS values.
`marginBlock` / `marginInline` / `paddingBlock` / `paddingInline` have
been available in all browsers since April 2021. `maxSize` and `minSize`
don't appear to be real CSS properties, and would be better suited to be
configured as custom shorthands in the stitches config.

this resolves an issue with the "splitBySpace" approach that does not
correctly handle `!important`